separating authOptions
